% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{assumptions.cmpoutput}
\alias{assumptions.cmpoutput}
\title{Get assumptions for parametric tests performed on output comparisons.}
\usage{
\method{assumptions}{cmpoutput}(cmpout)
}
\arguments{
\item{cmpout}{Object of class \code{cmpoutput}.}
}
\value{
Object of class \code{assumptions_cmpoutput} containing the
assumptions for parametric tests performed on an output comparisons
Basically a list containing the assumptions for the MANOVA (object of class
\code{\link{assumptions_manova}}) and univariate parametric tests for each
principal component (object of class \code{\link{assumptions_paruv}}).
}
\description{
Get assumptions for parametric tests performed on output comparisons (i.e.
from objects of class \code{\link{cmpoutput}}).
}
\examples{
NULL
}

