% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{assumptions.micomp}
\alias{assumptions.micomp}
\title{Get assumptions for parametric tests performed on each comparisons.}
\usage{
\method{assumptions}{micomp}(mcmp, ...)
}
\arguments{
\item{mcmp}{Object of class \code{\link{micomp}}.}

\item{...}{Currently ignored.}
}
\value{
Object of class \code{assumptions_micomp} containing the
assumptions for parametric tests performed for the multiple comparisons held
by the \code{mcmp} object. This object is a multi-dimensional list of
\code{assumptions_cmpoutput} objects. Rows are associated with individual
outputs, while columns are associated with separate comparisons.
}
\description{
Get assumptions for parametric tests performed on multiple comparisons (i.e.
from objects of class \code{\link{micomp}}).
}
\examples{
NULL
}

