% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{Get assumptions for parametric tests applied to the comparison of simulation
outputs}
\usage{
assumptions(obj, ...)
}
\arguments{
\item{obj}{Object from which to get the assumptions.}

\item{...}{Further arguments passed to concrete methods.}
}
\value{
Assumptions for parametric tests applied to the comparison of
simulation outputs.
}
\description{
Generic function to get the assumptions for parametric tests applied to the
comparison of simulation outputs.
}
\examples{
NULL
}

