% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions_manova}
\alias{assumptions_manova}
\title{Title}
\usage{
assumptions_manova(data, factors)
}
\arguments{
\item{data}{}

\item{factors}{}
}
\value{
to do
}
\description{
Title
}
\examples{
#' todo indeed
}

