% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions_parunivar}
\alias{assumptions_parunivar}
\title{Title}
\usage{
assumptions_parunivar(data, factors)
}
\arguments{
\item{data}{}

\item{factors}{}
}
\value{
to do
}
\description{
Title
}
\examples{
#' todo
}

