% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{micomp}
\alias{micomp}
\title{micomp...}
\usage{
micomp(outputs, nvars, ve, ..., concat = F)
}
\arguments{
\item{nvars}{}

\item{ve}{}

\item{nout}{}

\item{comps}{}
}
\value{
Some stuff
}
\description{
micomp...
}
\examples{
#' micomp()
}

