% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micompr-package.R
\docType{package}
\name{micompr-package}
\alias{micompr}
\alias{micompr-package}
\title{micompr: Multivariate Independent Comparison of Observations}
\description{
A procedure for comparing multivariate samples associated with different groups. It uses principal component analysis to convert multivariate observations into a set of linearly uncorrelated statistical measures, which are then compared using a number of statistical methods. The procedure is independent of the distributional properties of samples and automatically selects features that best explain their differences, avoiding manual selection of specific points or summary statistics. It is appropriate for comparing samples of time series, images, spectrometric measures or similar multivariate observations. This package is described in Fachada et al. (2016) \doi{10.32614/RJ-2016-055}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nunofachada/micompr}
  \item Report bugs at \url{https://github.com/nunofachada/micompr/issues}
}

}
\author{
\strong{Maintainer}: Nuno Fachada \email{faken@fakenmc.com} (\href{https://orcid.org/0000-0002-8487-5837}{ORCID})

}
\keyword{internal}
