% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micompr.R
\docType{package}
\name{micompr}
\alias{micompr}
\alias{micompr-package}
\title{micompr: multivariate independent comparison of observations}
\description{
micompr is an R package for independent comparison of multivariate
observations.
}
\note{
MIT License
}
\section{Usage}{


To do.
}
\references{
\href{http://arxiv.org/abs/1603.06907}{micompr: An R Package for
Multivariate Independent Comparison of Observations}, arXiv preprint.
}

