% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{plot.assumptions_manova}
\alias{plot.assumptions_manova}
\title{Title}
\usage{
\method{plot}{assumptions_manova}(asmnv, ...)
}
\arguments{
\item{asmnv}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

