% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{plot.assumptions_micomp}
\alias{plot.assumptions_micomp}
\title{Title}
\usage{
\method{plot}{assumptions_micomp}(micas, col = micomp:::plotcols(), ...)
}
\arguments{
\item{micas}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

