% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{plot.cmpoutput}
\alias{plot.cmpoutput}
\title{Title}
\usage{
\method{plot}{cmpoutput}(cmpout, col = c("blue", "red", "green", "gold",
  "violet", "cyan"), ...)
}
\arguments{
\item{cmpout}{}

\item{col}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo()
}

