% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{plot.cmpoutput_assumptions}
\alias{plot.cmpoutput_assumptions}
\title{Title}
\usage{
\method{plot}{cmpoutput_assumptions}(cmpoass, ...)
}
\arguments{
\item{cmpoass}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
NULL
}

