% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grpoutputs.R
\name{plot.grpoutputs}
\alias{plot.grpoutputs}
\title{Title}
\usage{
\method{plot}{grpoutputs}(go, col = micomp:::plotcols(), ...)
}
\arguments{
\item{go}{}

\item{col}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo()
}

