% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{plot.micomp}
\alias{plot.micomp}
\title{Title}
\usage{
\method{plot}{micomp}(mcmp, col = c("blue", "red", "green", "gold", "violet",
  "cyan"), ...)
}
\arguments{
\item{mcmp}{}

\item{col}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo()
}

