% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotcols}
\alias{plotcols}
\title{Default colors for plots in \code{micomp} package}
\usage{
plotcols()
}
\value{
Default colors for plots in \code{micomp} package.
}
\description{
Default colors for plots in \code{micomp} package.
}
\examples{
micompr:::plotcols()
# [1] "blue"   "red"    "green"  "gold"   "violet" "cyan"

}
\keyword{internal}

