% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pphpc_ok.R
\docType{data}
\name{pphpc_ok}
\alias{pphpc_ok}
\title{Data from two similar implementations of the PPHPC model}
\format{A \code{\link{grpoutputs}} object containing simulation output data
from 20 runs of the PPHPC model, 10 runs from each implementation. The model
has six outputs, but the object contains a seventh output corresponding to
the concatenation of the six outputs}
\source{
Runs are obtained from the NetLogo and Java (EX with 8 threads)
implementations of the PPHPC model available at
\url{https://github.com/fakenmc/pphpc}. The \code{config400v1.txt}
configuration was used in both cases.
}
\usage{
pphpc_ok
}
\description{
A dataset containing simulation output data from two implementations of the
PPHPC model.
}
\keyword{datasets}
