% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_manova}
\alias{print.assumptions_manova}
\title{Print information about the assumptions of the MANOVA test}
\usage{
\method{print}{assumptions_manova}(asmnv)
}
\arguments{
\item{asmnv}{Object of class \code{assumptions_manova}.}
}
\value{
The argument \code{asmnv}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_manova}, which
represent the assumptions of the MANOVA test performed on a comparison of
simulation output.
}
\examples{
NULL
}

