% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{print.assumptions_micomp}
\alias{print.assumptions_micomp}
\title{Print information about the assumptions concerning the parametric tests
performed on multiple comparisons of simulation output}
\usage{
\method{print}{assumptions_micomp}(micas, ...)
}
\arguments{
\item{micas}{Object of class \code{assumptions_micomp}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{micas}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_micomp}, which
represent the assumptions concerning the parametric tests performed on
multiple comparisons of simulation output.
}
\examples{
NULL
}

