% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_paruv}
\alias{print.assumptions_paruv}
\title{Title}
\usage{
\method{print}{assumptions_paruv}(aspuv)
}
\arguments{
\item{aspuv}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

