% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{print.cmpoutput}
\alias{print.cmpoutput}
\title{Print information about comparison of simulation output}
\usage{
\method{print}{cmpoutput}(cmpout)
}
\arguments{
\item{cmpout}{Object of class \code{cmpoutput}.}
}
\value{
The argument \code{cmpout}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{cmpoutput}.
}
\examples{
NULL
}

