% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grpoutputs.R
\name{print.grpoutputs}
\alias{print.grpoutputs}
\title{Print information about grouped outputs}
\usage{
\method{print}{grpoutputs}(go)
}
\arguments{
\item{go}{Object of class \code{grpoutputs}.}
}
\value{
The argument \code{go}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{grpoutputs}.
}
\examples{
NULL
}

