% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{print.micomp}
\alias{print.micomp}
\title{Print information about multiple comparisons of simulation output}
\usage{
\method{print}{micomp}(mcmp)
}
\arguments{
\item{mcmp}{Object of class \code{\link{micomp}}.}
}
\value{
The argument \code{mcmp}, invisibly, as for all \code{\link{print}}.
methods.
}
\description{
Print information about objects of class \code{\link{micomp}}.
}
\examples{
NULL
}

