% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp_latex.R
\name{pst}
\alias{pst}
\title{Concatenate strings without any separator characters}
\usage{
pst(...)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}
}
\value{
A character vector of the concatenated values without any separator
characters.
}
\description{
Concatenate strings without any separator characters.
}
\details{
This function simply calls \code{\link{paste}} with the \code{sep} and
\code{collapse} options set to \code{""}.
}
\examples{
micomp:::pst("a","b","c",c("a","b","c"))
# [1] "abcaabcbabcc"
}
\keyword{internal}

