% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp_latex.R
\name{pvalf}
\alias{pvalf}
\title{Format a \emph{p}-value for printing in a \code{LaTeX} table.}
\usage{
pvalf(pval)
}
\arguments{
\item{pval}{Numeric value between 0 and 1.}
}
\value{
A string representing the formatted \code{pval}.
}
\description{
Format a \emph{p}-value for printing in a \code{LaTeX} table. Requires the
\emph{ulem} \code{LaTeX} package for underlining the \emph{p}-values.
}
\examples{
pvalf(0.1)
# [1] "0.100"
pvalf(0.000001)
# [1] "\\\\uuline{1e-06}"
}
\keyword{internal}

