% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pvalnum}
\alias{pvalnum}
\title{Make sure \emph{p}-values are numeric}
\usage{
pvalnum(pvals)
}
\arguments{
\item{pvals}{Vector of \emph{p}-values which might not be fully numeric.}
}
\value{
A vector of fully numeric \emph{p}-values.
}
\description{
Make sure \emph{p}-values are numeric. Non-numeric values (e.g., "<0.001")
are converted to zero.
}
\examples{
micompr:::pvalnum(c("0.06", "0.9", "<0.0001", "0.3"))
# [1] 0.06 0.90 0.00 0.30

}
\keyword{internal}
