% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{summary.assumptions_cmpoutput}
\alias{summary.assumptions_cmpoutput}
\title{Summary method for the assumptions of parametric tests used in a comparison
of simulation outputs.}
\usage{
\method{summary}{assumptions_cmpoutput}(object, ...)
}
\arguments{
\item{object}{Object of class \code{assumptions_cmpoutput}.}

\item{...}{Currently ignored.}
}
\value{
A vector with the following \emph{p}-values:
\describe{
 \item{Royston(\emph{npcs,group})}{One per \emph{npcs x group} combination,
       yielded by the Royston test (\code{\link[MVN]{roystonTest}}) for the
       respective number of principal components (for a given explained
       variance) and group combination.}
 \item{BoxM(npcs)}{One per \emph{npcs}, yielded by Box's M test
       (\code{\link[biotools]{boxM}}).}
 \item{Shapiro-Wilk(\emph{group})}{One per group, yielded by the Shapiro-Wilk
       test (\code{\link{shapiro.test}}) for the first principal component of
       the respective group.}
 \item{Bartlett}{One row with the \emph{p}-value yielded by Bartlett's
       test (\code{\link{bartlett.test}}) for the first principal component.}
}
}
\description{
Summary method for objects of class \code{assumptions_cmpoutput}, which
contain the assumptions for the parametric tests used in a comparison of
simulation output.
}
\examples{
# TODO
}

