% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{summary.assumptions_micomp}
\alias{summary.assumptions_micomp}
\title{Summary method for the assumptions of parametric tests used in multiple
comparisons of simulation output}
\usage{
\method{summary}{assumptions_micomp}(object, ..., tnpcs = 1)
}
\arguments{
\item{object}{Object of class \code{assumptions_micomp}.}

\item{...}{Currently ignored.}

\item{tnpcs}{Number of principal components to summarize for the
\emph{t}-test.}
}
\value{
A multi-dimensional list of \code{assumptions_cmpoutput} summaries.
Rows are associated with individual outputs, while columns are associated
with separate comparisons.
}
\description{
Summary method for objects of class \code{assumptions_micomp}, which
contain the assumptions for the parametric tests used in multiple comparisons
of simulation output.
}
\examples{
# Create a micomp object, use provided dataset
mic <- micomp(5, 0.8,
              list(list(name = "NLOKvsJEXOK", grpout = pphpc_ok),
                   list(name = "NLOKvsJEXNOSHUFF", grpout = pphpc_noshuff)),
              concat = TRUE)

# Get the assumptions summary
sam <- summary(assumptions(mic))
}

