% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grpoutputs.R
\name{summary.grpoutputs}
\alias{summary.grpoutputs}
\title{Summary method for grouped outputs}
\usage{
\method{summary}{grpoutputs}(go)
}
\arguments{
\item{go}{Object of class \code{grpoutputs}.}
}
\value{
A list with the following components:
\describe{
 \item{output.dims}{Dimensions of each output, i.e. number of observations
          and number of variables (i.e. output length).}
 \item{group.sizes}{Number of output observations in each group.}
}
}
\description{
Summary method for objects of class \code{grpoutputs}.
}
\examples{
NULL
}

