% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp_latex.R
\name{toLatex.micomp}
\alias{toLatex.micomp}
\title{Title}
\usage{
\method{toLatex}{micomp}(mic, data.show = c("npcs", "mnvp", "parp", "nparp",
  "scoreplot"), table.placement = "ht", latex.environments = c("center"),
  booktabs = F, btalign = "l", col.width = F, digits = 3,
  pvalformat = pvalf,
  marks = c("mark=square*,mark options={color=red},mark size=0.8pt",
  "mark=*,mark size=0.6pt", "mark=o,mark size=0.7pt"), ...)
}
\arguments{
\item{mic}{}

\item{data.show}{}

\item{table.placement}{}

\item{latex.environments}{}

\item{booktabs}{}

\item{btalign}{}

\item{col.width}{}

\item{digits}{}

\item{pvalformat}{}

\item{marks}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

