% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{assumptions.cmpoutput}
\alias{assumptions.cmpoutput}
\title{Title}
\usage{
\method{assumptions}{cmpoutput}(obj, ...)
}
\arguments{
\item{obj}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

