% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{Parametric tests assumptions}
\usage{
assumptions(obj)
}
\arguments{
\item{obj}{Object from which to get the assumptions.}
}
\value{
Assumptions for parametric tests applied to the comparison of
simulation outputs.
}
\description{
Generic function to get the assumptions for parametric tests applied to the
comparison of simulation outputs.
}
\seealso{
\code{\link{assumptions.cmpoutput}},
\code{\link{assumptions.micomp}}
}

