% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{cmpoutput}
\alias{cmpoutput}
\title{Title}
\usage{
cmpoutput(name, ve, data, factors)
}
\arguments{
\item{name}{}

\item{ve}{}

\item{data}{}

\item{factors}{}
}
\value{
Compared outputs
}
\description{
Title
}
\examples{
#' micomp()
}

