% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grpoutputs.R
\name{grpoutputs}
\alias{grpoutputs}
\title{Group outputs.}
\usage{
grpoutputs(outputs, nvars, folders, files, lvls = NULL, concat = F)
}
\arguments{
\item{outputs}{}

\item{nvars}{}

\item{folders}{}

\item{files}{}

\item{lvls}{}

\item{concat}{}
}
\value{
Some stuff
}
\description{
Group outputs.
}
\examples{
#' micomp()
}

