% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micompr.R
\docType{package}
\name{micompr}
\alias{micompr}
\alias{micompr-package}
\title{micompr: model-independent comparison of simulation output}
\description{
micompr is an R package for model-independent comparison of simulation
output.
}
\note{
MIT License
}
\section{Usage}{


To do.
}
\references{
\href{http://arxiv.org/abs/1509.09174}{Model-independent
comparison of simulation output}, arXiv preprint.
}

