% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{plot.assumptions_manova}
\alias{plot.assumptions_manova}
\title{Plot \emph{p}-values for testing the multivariate normality assumptions of
the MANOVA test}
\usage{
\method{plot}{assumptions_manova}(asmnv, ...)
}
\arguments{
\item{asmnv}{Objects of class \code{\link{assumptions_manova}}.}

\item{...}{Extra options passed to \code{\link{plot.default}}.}
}
\value{
None.
}
\description{
Plot method for objects of class \code{\link{assumptions_manova}} which
presents a bar plot containing the \emph{p}-values produced by the Royston
multivariate normality test (\code{\link[MVN]{roystonTest}}) for each group
being compared.
}
\examples{
NULL
}

