% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{plot.assumptions_micomp}
\alias{plot.assumptions_micomp}
\title{Plot \emph{p}-values for testing the assumptions of the parametric tests used
in multiple output comparison}
\usage{
\method{plot}{assumptions_micomp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{assumptions_micomp}.}

\item{...}{Extra options passed to \code{\link{barplot}}. The \code{col}
parameter will defines the colors to use for different tests (and groups in
the case of normality tests).}
}
\value{
None.
}
\description{
Plot method for objects of class \code{assumptions_cmpoutput}
containing \emph{p}-values produced by testing the assumptions of the
parametric tests used for multiple output comparisons.
}
\details{
Several bar plots are presented, one for each comparison, showing the
\emph{p}-values yielded by the Shapiro-Wilk (\code{\link{shapiro.test}}) and
Royston tests (\code{\link[MVN]{roystonTest}}) for univariate and
multivariate normality of each group, respectively, and for the Bartlett
(\code{\link{bartlett.test}}) and Box's M (\code{\link[biotools]{boxM}}) for
testing homogeneity of variances and of covariance matrices, respectively.
\emph{P}-values are aggregated by output in each plot. Note that the
\emph{p-values} plotted for the Shapiro-Wilk and Bartlett tests correspond to
group observations along the first principal component.
}
\examples{
# Create a micomp object, use provided dataset
mic <- micomp(6, 0.65,
              list(list(name = "NLOKvsJEXOK", grpout = pphpc_ok),
                   list(name = "NLOKvsJEXNOSHUFF", grpout = pphpc_noshuff),
                   list(name = "NLOKvsJEXDIFF", grpout = pphpc_diff)))

# Plot the p-values of the statistic tests evaluating the assumptions of the
# comparisons performed in the mic object
plot(assumptions(mic))
}

