% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{plot.assumptions_paruv}
\alias{plot.assumptions_paruv}
\title{Title}
\usage{
\method{plot}{assumptions_paruv}(aspuv, extra = 0, ...)
}
\arguments{
\item{aspuv}{}

\item{extra}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

