% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{plot.cmpoutput}
\alias{plot.cmpoutput}
\title{Plot comparison of simulation output}
\usage{
\method{plot}{cmpoutput}(cmpout, col = micompr:::plotcols(), ...)
}
\arguments{
\item{cmpout}{Object of class \code{cmpoutput}.}

\item{col}{Vector of colors to use on observations of different groups
(scatter plot only).}

\item{...}{Extra options passed to \code{\link{plot.default}}.}
}
\value{
None.
}
\description{
Plot objects of class  \code{cmpoutput}.
}
\details{
This method produces four sub-plots, namely:
\itemize{
  \item Scatter plot containing the projection of output observations on the
        first two dimensions of the principal components space.
  \item Bar plot of the percentage of variance explain per principal
        component.
  \item Bar plot of \emph{p}-values for the parametric test for each
        principal component.
  \item Bar plot of \emph{p}-values for the non-parametric test for each
        principal component.
}
}
\examples{
NULL
}

