% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{plot.micomp}
\alias{plot.micomp}
\title{Plot projection of output observations on the first two dimensions of the
principal components space}
\usage{
\method{plot}{micomp}(mcmp, col = micomp:::plotcols(), ...)
}
\arguments{
\item{mcmp}{An object of class \code{\link{micomp}}.}

\item{col}{Vector of colors to use on observations of different groups.}

\item{...}{Extra options passed to \code{\link{plot.default}}.}
}
\value{
None.
}
\description{
For each comparison and output combination, draw a scatter plot containing
the projection of output observations on the first two dimensions of the
principal components space.
}
\examples{
NULL
}

