% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{print.assumptions_cmpoutput}
\alias{print.assumptions_cmpoutput}
\title{Print method for the assumptions of parametric tests used in a comparison
of an output}
\usage{
\method{print}{assumptions_cmpoutput}(x, ...)
}
\arguments{
\item{x}{Object of class \code{assumptions_cmpoutput}.}

\item{...}{Currently ignored.}
}
\value{
None.
}
\description{
Print method for objects of class \code{assumptions_cmpoutput}, which
contain the assumptions for the parametric tests used in a comparison of an
output.
}
\examples{

# Create a cmpoutput object from the provided datasets
cmp <- cmpoutput("All", c(0.7, 0.8, 0.9),
                 pphpc_diff$data[["All"]], pphpc_diff$obs_lvls)

}

