% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_manova}
\alias{print.assumptions_manova}
\title{Title}
\usage{
\method{print}{assumptions_manova}(asmnv)
}
\arguments{
\item{asmnv}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

