% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_manova}
\alias{print.assumptions_manova}
\title{Print information about the assumptions of the MANOVA test}
\usage{
\method{print}{assumptions_manova}(x, ...)
}
\arguments{
\item{x}{Object of class \code{assumptions_manova}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{asmnv}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_manova}, which
represent the assumptions of the MANOVA test performed on a comparison of
simulation output.
}
\examples{
# Print information concerning the assumptions of applying MANOVA to the iris
# data (i.e. multivariate normality of each group and homogeneity of
# covariance matrices)
assumptions_manova(iris[, 1:4], iris[, 5])

## Royston test (Multivariate Normality):
##         P-value 'setosa': 2.187653e-06
##         P-value 'versicolor': 0.0847746
##         P-value 'virginica': 0.06776605
## Box's M test (Homogeneity of Covariance Matrices):
##         P-value: 3.352034e-20
}

