% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_manova}
\alias{print.assumptions_manova}
\title{Print information about the assumptions of the MANOVA test}
\usage{
\method{print}{assumptions_manova}(x, ...)
}
\arguments{
\item{x}{Object of class \code{assumptions_manova}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_manova}, which
represent the assumptions of the MANOVA test performed on a comparison of
outputs.
}
\examples{

# Print information concerning the assumptions of applying MANOVA to the iris
# data (i.e. multivariate normality of each group and homogeneity of
# covariance matrices)
assumptions_manova(iris[, 1:4], iris[, 5])

}

