% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_paruv}
\alias{print.assumptions_paruv}
\title{Print information about the assumptions of the parametric test}
\usage{
\method{print}{assumptions_paruv}(aspuv)
}
\arguments{
\item{aspuv}{Object of class \code{assumptions_paruv}.}
}
\value{
The argument \code{aspuv}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_paruv}, which
represent the assumptions of the parametric test (i.e. either
\code{\link{t.test}} or \code{\link{aov}}) performed on a comparison of
simulation output.
}
\examples{
# Print information about the assumptions of applying ANOVA to each column
# (dependent variable) of the iris data (i.e. normality of each group and
# homogeneity of variances)
assumptions_paruv(iris[, 1:4], iris[, 5])

## Shapiro-Wilk test (Normality):
##         P-value(s) 'setosa':  0.4595132 0.2715264 0.05481147 8.658573e-07
##         P-value(s) 'versicolor':  0.464737 0.3379951 0.1584778 0.0272778
##         P-value(s) 'virginica':  0.2583147 0.180896 0.1097754 0.08695419
## Bartlett test (Homogeneity of Variances):
##         P-value(s):  0.0003345076 0.3515028 9.229038e-13 3.054784e-09
}

