% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_paruv}
\alias{print.assumptions_paruv}
\title{Print information about the assumptions of the parametric test}
\usage{
\method{print}{assumptions_paruv}(aspuv)
}
\arguments{
\item{aspuv}{Object of class \code{assumptions_paruv}.}
}
\value{
The argument \code{aspuv}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_paruv}, which
represent the assumptions of the parametric test (i.e. either
\code{\link{t.test}} or \code{\link{aov}}) performed on a comparison of
simulation output.
}
\examples{
NULL
}

