% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{print.assumptions_paruv}
\alias{print.assumptions_paruv}
\title{Print information about the assumptions of the parametric test}
\usage{
\method{print}{assumptions_paruv}(x, ...)
}
\arguments{
\item{x}{Object of class \code{assumptions_paruv}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{assumptions_paruv}, which
represent the assumptions of the parametric test (i.e. either
\code{\link[stats]{t.test}} or \code{\link[stats]{aov}}) performed on a
comparison of outputs.
}
\examples{

# Print information about the assumptions of applying ANOVA to each column
# (dependent variable) of the iris data (i.e. normality of each group and
# homogeneity of variances)
assumptions_paruv(iris[, 1:4], iris[, 5])

}
