% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{print.cmpoutput}
\alias{print.cmpoutput}
\title{Print information about comparison of simulation output}
\usage{
\method{print}{cmpoutput}(cmpout)
}
\arguments{
\item{cmpout}{Object of class \code{cmpoutput}.}
}
\value{
The argument \code{cmpout}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{cmpoutput}.
}
\examples{
# Comparing the fifth output of the pphpc_diff dataset, which contains
# simulation output data from two implementations of the PPHPC model executed
# with a different parameter.

cmpoutput("WolfPop", 0.7, pphpc_diff$data[[5]], pphpc_diff$factors)

## Output name: WolfPop
## Number of PCs which explain 70\% of variance: 2
## P-Value for MANOVA along 2 dimensions: 5.533087e-08
## P-Value for t-test (1st PC): 1.823478e-08
## P-Value for Mann-Whitney U test (1st PC): 1.082509e-05
}

