% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{print.cmpoutput}
\alias{print.cmpoutput}
\title{Print information about comparison of an output}
\usage{
\method{print}{cmpoutput}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cmpoutput}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
methods.
}
\description{
Print information about objects of class \code{cmpoutput}.
}
\examples{

# Comparing the fifth output of the pphpc_diff dataset, which contains
# simulation output data from two implementations of the PPHPC model executed
# with a different parameter.

cmpoutput("WolfPop", 0.7, pphpc_diff$data[[5]], pphpc_diff$obs_lvls)

}

