% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micomp.R
\name{print.micomp}
\alias{print.micomp}
\title{Print information about multiple comparisons of outputs}
\usage{
\method{print}{micomp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{micomp}}.}

\item{...}{Currently ignored.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}.
methods.
}
\description{
Print information about objects of class \code{\link{micomp}}.
}
\examples{
\donttest{
# A micomp object from package datasets (i.e. grpoutputs objects) directly

micomp(c("outA", "outB", "outC", "outD"), 0.9,
              list(list(name = "Comp1", grpout = pphpc_ok),
                   list(name = "Comp2", grpout = pphpc_noshuff),
                   list(name = "Comp3", grpout = pphpc_diff)))
}
}
