% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp_latex.R
\name{pvalf}
\alias{pvalf}
\title{Format p-values}
\usage{
pvalf(pval, params)
}
\arguments{
\item{pval}{Numeric \emph{p}-value to format (between 0 and 1).}

\item{params}{A list of method-dependent options.}
}
\value{
A string representing the formatted \emph{p}-value.
}
\description{
Generic function to format \emph{p}-values.
}
\seealso{
\code{\link{pvalf.default}}
}

