% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{summary.assumptions_micomp}
\alias{summary.assumptions_micomp}
\title{Title}
\usage{
\method{summary}{assumptions_micomp}(micas, ...)
}
\arguments{
\item{micas}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo
}

