% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp.R
\name{summary.assumptions_micomp}
\alias{summary.assumptions_micomp}
\title{Summary method for the assumptions of parametric tests used in multiple
comparisons of simulation output}
\usage{
\method{summary}{assumptions_micomp}(micas)
}
\arguments{
\item{micas}{Object of class \code{assumptions_micomp}.}
}
\value{
A list in which each component is associated with a distinct
comparison. Each component contains a data frame, in which columns represent
individual simulation outputs and rows have information about the assumptions
of the parametric tests used in each output. More specifically, each data
frame has rows with the following information:
\describe{
 \item{Royston(\emph{group})}{\emph{s} rows, one per group, with the
       \emph{p}-value yielded by the Royston test
       (\code{\link[MVN]{roystonTest}}) for the respective group.}
 \item{BoxM(Var.)}{One row with the \emph{p}-value yielded by Box's M test
       (\code{\link[biotools]{boxM}}).}
 \item{Shapiro-Wilk(\emph{group})}{\emph{s} rows, one per group, with the
       \emph{p}-value yielded by the Shapiro-Wilk test
       (\code{\link{shapiro.test}}) for the respective group.}
 \item{Bartlett(Var.)}{One row with the \emph{p}-value yielded by Bartlett's
       test (\code{\link{bartlett.test}}).}
}
}
\description{
Summary method for objects of class \code{assumptions_micomp}, which
contain the assumptions for the parametric tests used in multiple comparisons
of simulation output.
}
\examples{
# Create a micomp object, use provided dataset
mic <- micomp(5, 0.8,
              list(list(name = "NLOKvsJEXOK", grpout = pphpc_ok),
                   list(name = "NLOKvsJEXNOSHUFF", grpout = pphpc_noshuff)),
              concat = TRUE)

# Get the assumptions summary
sam <- summary(assumptions(mic))
}

