% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmpoutput.R
\name{summary.cmpoutput}
\alias{summary.cmpoutput}
\title{Title}
\usage{
\method{summary}{cmpoutput}(cmpout)
}
\arguments{
\item{cmpout}{}
}
\value{
todo
}
\description{
Title
}
\examples{
#' todo()
}

