% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micomp_latex.R
\name{toLatex.micomp}
\alias{toLatex.micomp}
\title{Title}
\usage{
\method{toLatex}{micomp}(mic, data.show = c("npcs", "mnvp", "parp", "nparp",
  "scoreplot"), table.placement = "ht", latex.environments = c("center"),
  booktabs = F, booktabs.cmalign = "l", caption = NULL, label = NULL,
  col.width = F, digits = 3, pvalformat = pvalf,
  scoreplot.marks = c("mark=square*,mark options={color=red},mark size=0.8pt",
  "mark=*,mark size=0.6pt", "mark=o,mark size=0.7pt"), scoreplot.scale = 6,
  scoreplot.before = "\\\\raisebox{-.5\\\\height}{\\\\resizebox {1.2cm} {1.2cm} {",
  scoreplot.after = "}}", ...)
}
\arguments{
\item{mic}{}

\item{data.show}{}

\item{table.placement}{}

\item{latex.environments}{}

\item{booktabs}{}

\item{booktabs.cmalign}{}

\item{caption}{}

\item{label}{}

\item{col.width}{}

\item{digits}{}

\item{pvalformat}{}

\item{scoreplot.marks}{}

\item{scoreplot.scale}{}

\item{scoreplot.before}{}

\item{scoreplot.after}{}

\item{...}{}
}
\value{
todo
}
\description{
Title
}
\examples{
NULL
}

